requirements.txt
anthropic>=0.40.0,<1.0
pydantic>=2.0,<3
python-dotenv>=1.0
langfuse>=3.0
pytest>=8.0
requests>=2.31
fastapi>=0.100
uvicorn>=0.20
google-auth>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
google-api-python-client>=2.0
ruff>=0.4
